backend/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
PyPDF2
python-docx
python-dotenv
langchain
langchain-google-genai
pytest
pytest-asyncio
pytest-mock
pytest-cov